Type defining intras checking
I have written the initial code (no debugging yet) for checking each of
the least abundant defining intras in the types and getting rid of them
if they are not common enough. This is due to the fact that we don't
want types defined by very rare intras as there is a chance that these
intras will not be sequcened in the average amplicon library even if the
intra would be present with infinite reads. We will have to debug this
and work out how the level of support parameters will be working into
the config file. The next major code that needs implementing is code
that will check the final type assignments for correct ratios. For
example, if a type is a codom and the proportions are fairly close to
that normally and we find the intras in a given sample but the ratios
are way off then we will not assign this type.